Fix map and sequence end locations and overlapping
`Event(Mapping|Sequence)End` events have a location of the start of the
next event. This is too large, if there's whitespace or comments between
the actual end. To address this, we track the (end) location of the last
thing we parsed within the mapping or sequence and consider that the end
of the mapping or sequence itself.

To accomplish this, we refactor `Marked` to make testing clearer and
then beef up the `DecodeSpec` to verify marks fully throughout a more
complex document.

Before doing this, I thought the handling of overlapping `Replace`s was
broken, so I refactored that as well. I like the result, so let's keep
it.