Add markedJSONPath, prepend in error messages
This is wildly tricky.

When aeson does it's parsing in `Parser` it knows the current 'JSONPath'
as part of state. Failure can then be handled in an overall way with the
current set of `JSONPath` values (such as `$`, `$.foo`, `$.foo.0`) as
part of running the `Parser`, finally rendering (e.g.) `Error in
$['foo'][0]` at the end.

This is what we're recreating.

However, we don't have `Parser` (just `Either String`). Further, we
parse the events to a `Marked Value` first, completely, then parse from
that later.

This causes two problems that we deal with separately:

1. We don't have `JSONPath` where we need it (when handling `Marked
   Value`s). To solve that, we added `markedJSONPath`. This is a
   breaking change.

2. As we prepend each of `$`, `$.foo`, and `$.foo.0` to our errors we
   never get to see them all at once (where one may like to, say, "take
   longest"), we have to prepend then one at a time, which means we need
   to avoid producing, `$.$.foo.$.foo.0`. To solve this, we only prepend
   the last element of any given path, assuming leading elements will
   then get prepended on their own. Lastly, we use the `[` character in
   the thing we're prepending onto, to know if we should add the `: `
   separator or not.

Phew.